web_requirements.txt
# Flask Web Server Requirements
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Your existing analysis.py dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0

# Additional web server utilities
python-multipart>=0.0.6
requests>=2.28.0